Including path awareness to support Path validation
This commit includes many changes. The main features it brings are
the following.

Bringing network awareness

The stack currently has only information about the peer's IP and
port. This is not sufficient for, e.g., a client that wants to use
two different network interfaces (or even two different source
ports) over a given connection. This commit brings full 4-tuple
awareness to the stack, i.e. both the peer and local addresses.
This change breaks some existing public APIs, such as `SendInfo`,
`RecvInfo`, `connect()` and `accept()`, but applications can be
easily updated as the local address can be typically retrieved
using `socket.local_addr().unwrap()`. In addition, this commit
offers to the application the ability to consider only a specific
4-tuple using the `path_send()` method (implementers might also be
interested in the `sendable_paths()` method).

Including the notion of Path

Until now, the stack assumed a single path for the connection. This
commit enables a connection to use several network paths and keep
track of packets sent on them thanks to the `Path` structure
(located in its dedicated `path` module). Note that the `Recovery`
structure is now relative to a path and not on the whole connection
anymore. This enables the stack to easily maintain sent packets,
RTT estimates, congestion control state,... on each existing path.
All these network paths are handled by a dedicated structure, the
`PathManager`, also located in the `path` module.

Path validation support

The stack is now able to probe network paths in order to validate
them through PATH_CHALLENGE/PATH_RESPONSE exchange. Such challenge
process can be triggered by the application thanks to the
`probe_path()` method. While the client can probe new 4-tuples, the
server first needs to see a peer's packet originating from the
4-tuple before probing it. Several new `PathEvent` events returned
by the `poll()` method (`New`, `Validated`, `FailedValidation`,...)
enables the application to handle such cases.