Connection migration support
quiche now supports connection migration, either passively as a
server (raising a `PathEvent::PeerMigrated` event to the app), or
actively as a client through the `migrate()` API. Note that after
a migration, the connection might lose connectivity, e.g. because
the client uses a non-working 4-tuple or because there was some
form of address spoofing attack seen as a migration by the server.
quiche includes a fall-back support when the newly migrated path is
detected as non-working.

Note that this commit includes a way for the host to control how it
handles migration without having spare DCID. By default, it reuses
the DCID of the previous active path. The application can prevent
this behavior and stall until having a new DCID from its peer by
calling `set_disable_dcid_reuse()` on the `Config`.

This commit also includes updated examples to show different usages
of the new provided API. See `quiche-client` and `quiche-server`.